Clean up celltypist shell script
Clean up singularity use - no longer needs to be explicit since
everything runs in the container. Also clean up the Gateway utility
directory, as it is more basic and clean now as well. It is provided
on the compute machine and the app is told where it is with a
parameter from the Gateway.